pyproject.toml
[project]
name = "snap-ai"
version = "0.1.0"
description = "A code sample for a Document AI / IDP system that combines traditional OCR + ML with Generative AI (GPT-4) using LangChain for intelligent document understanding (e.g., contracts, invoices, forms)."
authors = [
    {name = "Callis Ezenwaka", email = "callisezenwaka@outlook.com"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.2",
    "uvicorn[standard]==0.24.0",
    "python-multipart>=0.0.18",
    "jinja2==3.1.2",
    "python-dotenv==1.0.0",
    "openai==1.3.8",
    "langchain==0.0.350",
    "scikit-learn==1.3.2",
    "pypdf==5.8.0",
    "pillow==10.1.0",
    "pytesseract==0.3.10",
    "requests==2.31.0",
    "transformers==4.35.2",
    "torch==2.7.1",
    "starlette>=0.40.0",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["snap_ai"]
include-package-data = true

[tool.setuptools.package-data]
"*" = ["../templates/*", "../static/*"]